+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add a new benchmark matching option "-m exact" to match the benchmark name
+  exactly.
 * Write data to CSV file in quick mode too.
 * Fix the CSV file header to match with the data rows for the `--csvraw` case.